requirements.txt
cleverhans==3.0.1
configparser==3.7.4
imageio==2.5.0
matplotlib==2.2.4
numpy==1.16.2
opencv-contrib-python==4.1.0.25
pandas==0.24.2
Pillow==6.0.0
scikit-image==0.14.2
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
tensorboard==1.13.1
tensorflow-gpu==1.13.1
tqdm==4.32.1
